Harden the full-range GPU run: fix float32 promotion, NaN safety, checkpointing
Two independent review passes on the full-range/GPU path found real bugs that
would have wasted the 5-7h Kaggle run. Fixes + a robustness layer so a crash or
timeout never costs more than the board in flight:

Solver (batched_gpu.py)
- CRITICAL: _chance allocated float64 accumulators (missing dtype=self.dtype),
  silently promoting the whole "float32" run to float64 -> ~32x slower on a T4
  (FP64-crippled) -> Kaggle kills it at the session limit = the exact "no output"
  failure seen before. Also risked a mixed-dtype scatter_add crash. Now typed.
- Belt-and-suspenders: _update casts regrets back to self.dtype.
- No test exercised float32 or streets=3, which is why it slipped through.

Explanations (explanations.py)
- explain() called round(100*freq) which RAISES on NaN/inf (ValueError/Overflow),
  and runs inside the per-board loop -> one bad hand killed the entire run. Now
  the frequency line is skipped when any value is non-finite.

Pipeline (content_yield.py)
- Per-board checkpointing: solve -> validate -> write boards/board_XX.json;
  refresh records.json + yield_report.json after EVERY board (partial output is
  always present/downloadable); a re-run skips completed boards (resume).
- Drop only individual non-finite (NaN/inf) records so a float32 blow-up on one
  hand neither discards the board nor leaks into the signed pack (which would
  also have broken the final signature via SQLite NaN->NULL coercion).
- One crashing board is logged (board_XX.ERROR.txt) and the run continues.
- Fix deflated yield projection: measure the true range and set scale=1.0 when
  --n >= range (was 250/400=0.625, understating the target check ~1.6x).
- New CLI: --fresh, --aggregate-only.

Tests
- test_gpu_float32.py: full street-3 tree (+raise) on the numpy backend at
  float32 must keep all regret/strategy tables float32 and report finite EVs.

Notebook (kaggle_content_yield.ipynb v5)
- Re-embeds the fixed source; documents the checkpointing/partial-safe behavior;
  build cell prints boards_completed/missing and is partial-safe.

All 55 tests pass. Round-trip verified: extracting the embedded notebook zip and
running solve -> records -> build -> verify reproduces hash_ok + signature_ok.

+++ src/pokertrainer/solver/batched_gpu.py
@@ -172,7 +172,7 @@ class BatchedGPUCFR:
                                  child_ro, child_ri, path)
         uo_c = uo_c * self.o_alive[pc_d]
         ui_c = ui_c * self.i_alive[pc_d]
-        UO = xp.zeros((C, self.no)); UI = xp.zeros((C, self.ni))
+        UO = xp.zeros((C, self.no), dtype=self.dtype); UI = xp.zeros((C, self.ni), dtype=self.dtype)
         self.scatter_add(UO, pk_d, uo_c)
         self.scatter_add(UI, pk_d, ui_c)
         return UO / denom, UI / denom
@@ -304,7 +304,9 @@ class BatchedGPUCFR:
         if self._eval:
             return
         base = (s * u).sum(axis=2, keepdims=True)
-        self.R[key] = self.xp.maximum(self.R[key] + (u - base), 0.0)
+        # keep regrets in self.dtype — a stray float64 here would silently promote
+        # the whole float32 run and cripple throughput on FP64-limited GPUs (T4).
+        self.R[key] = self.xp.maximum(self.R[key] + (u - base), 0.0).astype(self.dtype, copy=False)
         self.S[key] += self._t * reach[:, :, None] * s
 
     def _eval_capture(self):
